// src/api/models/createInsightLogReqDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { CreateInsightLogReqDTOCategory } from './createInsightLogReqDTOCategory';

export interface CreateInsightLogReqDTO {
  /** @maxLength 20 */
  title: string;
  /** @maxLength 250 */
  description: string;
  category: CreateInsightLogReqDTOCategory;
  activityIds: unknown[][];
}