// src/api/models/insightControllerCreateActivityTag200.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityNameResDTO } from './activityNameResDTO';
import type { InsightControllerCreateActivityTag200Error } from './insightControllerCreateActivityTag200Error';

export type InsightControllerCreateActivityTag200 = {
  timestamp?: string;
  isSuccess?: boolean;
  /** @nullable */
  error?: InsightControllerCreateActivityTag200Error;
  result?: ActivityNameResDTO;
};