// src/api/models/insightControllerCreateLog200.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { InsightControllerCreateLog200Error } from './insightControllerCreateLog200Error';
import type { InsightLogResDTO } from './insightLogResDTO';

export type InsightControllerCreateLog200 = {
  timestamp?: string;
  isSuccess?: boolean;
  /** @nullable */
  error?: InsightControllerCreateLog200Error;
  result?: InsightLogResDTO;
};