// src/api/models/insightControllerGetActivityTags200.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityNameResDTO } from './activityNameResDTO';
import type { InsightControllerGetActivityTags200Error } from './insightControllerGetActivityTags200Error';

export type InsightControllerGetActivityTags200 = {
  timestamp?: string;
  isSuccess?: boolean;
  /** @nullable */
  error?: InsightControllerGetActivityTags200Error;
  result?: ActivityNameResDTO[];
};