// src/api/models/insightControllerGetLogs200.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { InsightControllerGetLogs200Error } from './insightControllerGetLogs200Error';
import type { InsightLogResDTO } from './insightLogResDTO';

export type InsightControllerGetLogs200 = {
  timestamp?: string;
  isSuccess?: boolean;
  /** @nullable */
  error?: InsightControllerGetLogs200Error;
  result?: InsightLogResDTO[];
};